requirements.txt
click==8.3.2
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
qrcode==8.2
Pillow==12.2.0
Werkzeug==3.1.8
gunicorn==23.0.0
svglib==1.5.1
reportlab==4.3.1